refactor(prisma): clean-break legacy rate breakdown and Decimal columns [ai-gen]
Remove rateBreakdownJson legacy flag and migrate tax rates plus space area to Int storage. Drop Prisma $extends/basePrisma in favor of a single client singleton per official Prisma + Better Auth guidance.

+++ src/app/(admin)/admin/(dashboard)/_shared/actions/settings/schemas/discount.ts
@@ -17,8 +17,8 @@ import { settingsExpectedUpdatedAtSchema } from "./form-schema-helpers";
 export const taxDisplayModeSchema = z.enum(TaxDisplayMode);
 
 export const taxSettingsSchema = z.object({
-  taxStandardRate: z.coerce.number().min(0).max(100),
-  taxReducedRate: z.coerce.number().min(0).max(100),
+  taxStandardRate: z.coerce.number().int().min(0).max(100),
+  taxReducedRate: z.coerce.number().int().min(0).max(100),
   taxDisplayModePublic: taxDisplayModeSchema,
   expectedUpdatedAt: settingsExpectedUpdatedAtSchema,
 });
